+++ src/main/kotlin/app/revanced/patches/Index.kt
@@ -1,9 +1,10 @@
 package app.revanced.patches
 
 import app.revanced.patcher.patch.Patch
-import app.revanced.patches.ads.VideoAds
-import app.revanced.patches.layouts.CreateButtonRemover
-import app.revanced.patches.layouts.MinimizedPlayback
+import app.revanced.patches.ad.VideoAds
+import app.revanced.patches.interaction.EnableSeekbarTapping
+import app.revanced.patches.layout.*
+import kotlin.reflect.KClass
 
 /**
  * Index contains all the patches and should be imported when using this library.
@@ -14,9 +15,13 @@ object Index {
      * Array of patches.
      * New patches should be added to the array.
      */
-    val patches: Array<() -> Patch> = arrayOf(
-        ::VideoAds,
-        ::MinimizedPlayback,
-        ::CreateButtonRemover
+    val patches: Array<KClass<out Patch>> = arrayOf(
+        VideoAds::class,
+        MinimizedPlayback::class,
+        CreateButtonRemover::class,
+        HideReels::class,
+        HideSuggestions::class,
+        OldQualityLayout::class,
+        EnableSeekbarTapping::class
     )
 }